Make Workspace.cs abstract making room for GitWorkspace.cs
